Create README.md
I am new to Github and to the IT field. I’m open to meeting/collaborate others on here and develop my skills and knowledge. I have no other expectation but to succeed in whatever I put my mind to. Transitioning from Professional boxer to IT is a huge leap, and I am ready to get to work and have fun.